refactor: major refactor of audio player
The audio stuff has been moved into its own submodule, with the idea being that it can be enable or disabled in a future commit. The code has been generalized into an audio processor which can be used for synths or to render out the audio into a vec of f32s which can later be exported. The new audio player builds on top of the audio processor. There is also a new synth trait to support other types of synths apart from the wavetable synth in the future.