refactor: rework intervals and scales
The interval and scale modules have been re-enabled and reworked. Intervals are now a trait and the old values were moved to a specific implementation of Interval called TwelveToneInterval. Scales now only work with TwelveToneIntervals as well. Furthermore there is a new example of creating a seven tone equal temperament system.