refactor: clean up notes, intervals and minor improvements
Notes no longer require a tuning to be specified and only hold pitch class and octave now. TwelveToneInterval was renamed to StandardInterval because they may be used in other pitch class systems, and EqualTemperament and PythagoreanTuning were reverted back to calculating frequencies of notes instead of just intervals, which would have been a nightmare to handle.