refactor: rewrite chords and promote easy use of parameters
These changes reimplement chords so that they are two different structures, GenericChords which hold a PitchClass type, semitones and an inversion, and NoteChords that are bound to physical notes. These structures implement a macro to avoid code duplication and also allow TryFroms in their parameters, allowing users to pass many compatible types, such as strings.