refactor: universal conversions + scale & interval statics
These changes add conversions that allow the PitchClass, Note, Scale, Interval and Chord structures to be converted between each other under most contexts. The interval and scale classes have also been reworked, removing traits in favour of just having a new() function that allows the user to add new intervals or scales if they really desire to do so. The lazy_static library has also been added to allow dynamic types within these structs, and a mutex was created to have a global storage of all of the intervals that have been created in a hashmap.