refactor: smarter pointers and better types
Smarter pointers have been implemented for SynthRc (which is now known as SynthRef) and the tuning field of AudioPlayer, allowing these structures to be used in a thread safe manner. Redundant types such as CompleteChordError and others have been removed as well, and general improvements have been applied. There is also a new CONTRIBUTING.md file with guidelines that should be followed if people decide to open pull requests or suggest changes.